requirements.txt
Django>=6.0,<7.0
Pillow>=10.0
